% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coerce-WHE}
\alias{coerce-WHE}
\alias{coerce,WithinHostExperiment,data.frame-method}
\alias{coerce,WithinHostExperiment,VRanges-method}
\alias{coerce,VRanges,WithinHostExperiment-method}
\title{Coerce WithinHostExperiment to data.frame or VRanges}
\value{
A \code{data.frame} or \code{\link[VariantAnnotation]{VRanges}}
  object containing the variant data from the
  \code{WithinHostExperiment} object.
}
\description{
Coerce WithinHostExperiment to data.frame or VRanges
}
\examples{
data(example_whe)
df <- as(example_whe, "data.frame")
head(df)

vr <- as(example_whe, "VRanges")
vr
}
