% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{compareDiversity}
\alias{compareDiversity}
\title{Compare diversity between sample groups}
\usage{
compareDiversity(whe, groupBy, index = "shannon", genomeLength = NULL)
}
\arguments{
\item{whe}{A \code{\link{WithinHostExperiment}}.}

\item{groupBy}{Character. Column name in \code{colData} to group by.}

\item{index}{Character. Which index to compare (default: "shannon").}

\item{genomeLength}{Integer (optional). Required if index is "pi"
or "watterson".}
}
\value{
A \code{DataFrame} with test results.
}
\description{
Compare diversity between sample groups
}
\examples{
vcf <- system.file("extdata", "test_donor.vcf",
    package = "WithinHostExperiment")
whe <- readWithinHost(vcf,
    colData = S4Vectors::DataFrame(sample_id = "donor"),
    caller = "ivar")
## Single-group, returns diversity table:
compareDiversity(whe, groupBy = "sample_id", index = "shannon")
}
