% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottleneck-exact.R
\name{.bootstrapCI}
\alias{.bootstrapCI}
\title{Bootstrap confidence interval by resampling variant sites}
\usage{
.bootstrapCI(nboot, maxNb, n_var, method, ...)
}
\description{
Bootstrap confidence interval by resampling variant sites
}
\keyword{internal}
