% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottleneck-exact.R
\name{.extractPairData}
\alias{.extractPairData}
\title{Extract donor/recipient data for a pair}
\usage{
.extractPairData(whe, pairId, threshold, need_counts = FALSE)
}
\description{
Extract donor/recipient data for a pair
}
\keyword{internal}
