% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottleneck-exact.R
\name{.finalizeMLE}
\alias{.finalizeMLE}
\title{Finalize MLE from log-likelihood vector}
\usage{
.finalizeMLE(ll_vals, maxNb, nboot, n_var, pairId, method, ...)
}
\description{
Finalize MLE from log-likelihood vector
}
\keyword{internal}
