% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottleneck-exact.R
\name{.logBetaBinomPMF}
\alias{.logBetaBinomPMF}
\title{Log beta-binomial PMF: log BetaBinom(k | n, alpha, beta)}
\usage{
.logBetaBinomPMF(k, n, alpha, beta_p)
}
\description{
Log beta-binomial PMF: log BetaBinom(k | n, alpha, beta)
}
\keyword{internal}
