% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.makeWHE}
\alias{.makeWHE}
\title{Internal constructor from pre-validated components}
\usage{
.makeWHE(
  rowRanges,
  assays,
  colData,
  callerName = NA_character_,
  metadata = list()
)
}
\arguments{
\item{rowRanges}{\code{GRanges}. Variant positions.}

\item{assays}{Named list of matrices.}

\item{colData}{\code{DataFrame}. Sample metadata.}

\item{callerName}{Character. Caller name for metadata.}

\item{metadata}{Named list. Additional metadata.}
}
\value{
A \code{WithinHostExperiment}.
}
\description{
Internal constructor from pre-validated components
}
\keyword{internal}
