% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ISNVFilter}
\alias{ISNVFilter}
\title{Create an ISNVFilter object}
\usage{
ISNVFilter(
  minDepth = 100L,
  minFreq = 0.03,
  maxFreq = 0.5,
  minAltReads = 10L,
  maxStrandBias = 10,
  minBaseQual = 20L,
  minMapQual = 20L,
  replicateConc = FALSE
)
}
\arguments{
\item{minDepth}{Integer. Minimum read depth (default: 100).}

\item{minFreq}{Numeric. Minimum alt allele frequency (default: 0.03).}

\item{maxFreq}{Numeric. Maximum alt allele frequency (default: 0.50).}

\item{minAltReads}{Integer. Minimum alt allele reads (default: 10).}

\item{maxStrandBias}{Numeric. Max strand bias fold-diff (default: 10).}

\item{minBaseQual}{Integer. Minimum base quality (default: 20).}

\item{minMapQual}{Integer. Minimum mapping quality (default: 20).}

\item{replicateConc}{Logical. Require replicate concordance
(default: FALSE).}
}
\value{
An \code{\link{ISNVFilter}} object.
}
\description{
Constructs a quality control filter configuration for use with
\code{\link{flagISNV}}.
}
\examples{
ISNVFilter()
ISNVFilter(minDepth = 500L, minFreq = 0.05)
}
