% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{qcLog}
\alias{qcLog}
\alias{qcFilters}
\alias{passedISNV}
\alias{transmissionPairs}
\alias{qcLog<-}
\alias{qcFilters<-}
\alias{qcLog,WithinHostExperiment-method}
\alias{qcFilters,WithinHostExperiment-method}
\alias{qcLog<-,WithinHostExperiment-method}
\alias{qcFilters<-,WithinHostExperiment-method}
\alias{passedISNV,WithinHostExperiment-method}
\alias{transmissionPairs,WithinHostExperiment-method}
\title{Access the QC audit log}
\usage{
qcLog(x)

qcFilters(x)

passedISNV(x)

transmissionPairs(x)

qcLog(x) <- value

qcFilters(x) <- value

\S4method{qcLog}{WithinHostExperiment}(x)

\S4method{qcFilters}{WithinHostExperiment}(x)

\S4method{qcLog}{WithinHostExperiment}(x) <- value

\S4method{qcFilters}{WithinHostExperiment}(x) <- value

\S4method{passedISNV}{WithinHostExperiment}(x)

\S4method{transmissionPairs}{WithinHostExperiment}(x)
}
\arguments{
\item{x}{A \code{\link{WithinHostExperiment}} object.}

\item{value}{A list replacement value.}
}
\value{
A \code{DataFrame} with columns: step, parameter, value,
  n_flagged, n_passed, timestamp.

A list of \code{\link{ISNVFilter}} objects.

A \code{\link{WithinHostExperiment}} (subset of rows).

A \code{DataFrame} with columns: donor (sample_id),
  recipient (sample_id), pair_id.

The modified \code{\link{WithinHostExperiment}}.

The modified \code{\link{WithinHostExperiment}}.
}
\description{
Returns the quality control log recording all filtering steps
applied to a \code{\link{WithinHostExperiment}}.

Returns a subset of the \code{\link{WithinHostExperiment}} containing
only variant sites where at least one sample passed QC.

Identifies donor-recipient pairs based on \code{role} and
\code{pair_id} columns in \code{colData}.
}
\examples{
gr <- GenomicRanges::GRanges("seg", IRanges::IRanges(1:3, width = 1))
whe <- WithinHostExperiment(
    assays = list(altFreq = matrix(runif(3), ncol = 1)),
    rowRanges = gr,
    colData = S4Vectors::DataFrame(sample_id = "S1"))
qcLog(whe)
gr <- GenomicRanges::GRanges("seg", IRanges::IRanges(1:3, width = 1))
whe <- WithinHostExperiment(
    assays = list(altFreq = matrix(runif(3), ncol = 1)),
    rowRanges = gr,
    colData = S4Vectors::DataFrame(sample_id = "S1"))
passedISNV(whe)
}
