% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{combineRows}
\alias{combineRows}
\alias{combineCols}
\title{Combine WithinHostExperiment objects by rows or columns}
\usage{
combineRows(...)

combineCols(...)
}
\arguments{
\item{...}{\code{\link{WithinHostExperiment}} objects to combine.}
}
\value{
A \code{\link{WithinHostExperiment}}.
}
\description{
\code{combineRows} combines objects vertically (adding variant
sites); all objects must have the same samples (columns).
\code{combineCols} combines objects horizontally (adding
samples); all objects must have the same variant sites (rows).
QC logs and filters from all objects are merged.
}
\examples{
data(example_whe)
whe1 <- example_whe[1:10, ]
whe2 <- example_whe[11:nrow(example_whe), ]
merged <- combineRows(whe1, whe2)
nrow(merged) == nrow(example_whe)
data(example_whe)
whe_d <- example_whe[, 1]
whe_r <- example_whe[, 2]
merged <- combineCols(whe_d, whe_r)
ncol(merged)
}
